requirements.txt
flask
Biopython